requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.7
django-computed-property==0.3.0
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
pokebase==1.3.0
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
